@@ -3,24 +3,57 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 32
 
     defaultConfig {
         applicationId "newpointer.com.br.newpointerpedido"
-        minSdkVersion 14
-        targetSdkVersion 28
-        versionCode 42
-        versionName "4.02"
+        minSdkVersion 19
+        targetSdkVersion 32
+        versionCode 52
+        versionName "5.2"
         multiDexEnabled true
     }
+
+    compileOptions {
+        // Flag to enable support for the new language APIs
+        coreLibraryDesugaringEnabled true
+        // Sets Java compatibility to Java 8
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    /** signing Options **/
+    signingConfigs {
+        keyStoreInfo {
+            storeFile file("..\\keystore\\np_pedido_jks.jks")
+            storePassword "qpalzm"
+            keyPassword "qpalzm"
+            keyAlias "np_pedido"
+        }
+    }
+
     buildTypes {
         release {
             minifyEnabled false
+            shrinkResources false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            signingConfig signingConfigs.keyStoreInfo
         }
     }
+
     dexOptions {
         javaMaxHeapSize "4g"
+    }
+
+    lintOptions {
+        lintConfig file("lint.xml")
+        baseline file("lint-baseline.xml")
+        ignoreWarnings false
+        warningsAsErrors true
+        abortOnError true
+        checkReleaseBuilds false
+        showAll true
+        textOutput "stdout"
     }
 }
 
@@ -30,15 +63,23 @@ dependencies {
     implementation files('libs/Jaybird2_2_4.jar')
     implementation files('libs/android-viewbadger.jar')
 
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
     implementation 'com.google.android.gms:play-services:12.0.1'
-    implementation 'com.android.support:multidex:1.0.3'
-    implementation 'com.edwardvanraak:MaterialBarcodeScanner:0.0.6-ALPHA'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+//    implementation 'com.edwardvanraak:MaterialBarcodeScanner:0.0.6-ALPHA'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     testImplementation 'junit:junit:4.12'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+
+    implementation "androidx.activity:activity:1.4.0"
+    implementation "androidx.fragment:fragment:1.4.1"
+
+    implementation('com.journeyapps:zxing-android-embedded:4.3.0') { transitive = false }
+    implementation 'com.google.zxing:core:3.3.0'
+    coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:1.1.5'
+    implementation "androidx.multidex:multidex:2.0.1"
+
 }
 repositories {
     mavenCentral()
